AI: extract pathfinding special actions and rules to separate files
