+++ AI/VCAI/Pathfinding/Actions/BattleAction.cpp
@@ -1,0 +1,21 @@
+/*
+* BattleAction.cpp, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
+*/
+
+#include "StdInc.h"
+#include "../../Goals/VisitTile.h"
+#include "BattleAction.h"
+
+namespace AIPathfinding
+{
+	Goals::TSubgoal BattleAction::whatToDo(const HeroPtr & hero) const
+	{
+		return Goals::sptr(Goals::VisitTile(targetTile).sethero(hero));
+	}
+}
