+++ AI/VCAI/VCAI.vcxproj
@@ -168,10 +168,20 @@
     <ClCompile Include="Goals\Win.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="MapObjectsEvaluator.cpp" />
+    <ClInclude Include="Pathfinding\Actions\BattleAction.h" />
+    <ClInclude Include="Pathfinding\Actions\TownPortalAction.h" />
+    <ClCompile Include="Pathfinding\Actions\BattleAction.cpp" />
+    <ClCompile Include="Pathfinding\Actions\TownPortalAction.cpp" />
+    <ClInclude Include="Pathfinding\Actions\BoatActions.h" />
+    <ClCompile Include="Pathfinding\Actions\BoatActions.cpp" />
     <ClCompile Include="Pathfinding\AINodeStorage.cpp" />
     <ClCompile Include="Pathfinding\AIPathfinder.cpp" />
     <ClCompile Include="Pathfinding\AIPathfinderConfig.cpp" />
     <ClCompile Include="Pathfinding\PathfindingManager.cpp" />
+    <ClCompile Include="Pathfinding\Rules\AILayerTransitionRule.cpp" />
+    <ClCompile Include="Pathfinding\Rules\AIMovementAfterDestinationRule.cpp" />
+    <ClCompile Include="Pathfinding\Rules\AIMovementToDestinationRule.cpp" />
+    <ClCompile Include="Pathfinding\Rules\AIPreviousNodeRule.cpp" />
     <ClCompile Include="ResourceManager.cpp" />
     <ClCompile Include="SectorMap.cpp" />
     <ClCompile Include="StdInc.cpp">
@@ -214,10 +224,15 @@
     <ClInclude Include="Goals\VisitTile.h" />
     <ClInclude Include="Goals\Win.h" />
     <ClInclude Include="MapObjectsEvaluator.h" />
+    <ClInclude Include="Pathfinding\Actions\ISpecialAction.h" />
     <ClInclude Include="Pathfinding\AINodeStorage.h" />
     <ClInclude Include="Pathfinding\AIPathfinder.h" />
     <ClInclude Include="Pathfinding\AIPathfinderConfig.h" />
     <ClInclude Include="Pathfinding\PathfindingManager.h" />
+    <ClInclude Include="Pathfinding\Rules\AILayerTransitionRule.h" />
+    <ClInclude Include="Pathfinding\Rules\AIMovementAfterDestinationRule.h" />
+    <ClInclude Include="Pathfinding\Rules\AIMovementToDestinationRule.h" />
+    <ClInclude Include="Pathfinding\Rules\AIPreviousNodeRule.h" />
     <ClInclude Include="ResourceManager.h" />
     <ClInclude Include="SectorMap.h" />
     <ClInclude Include="StdInc.h" />
